Leave Room
- DB에서 룸 나가기 정보가 반영되지 않았던 점 수정
- Client 방나가기 기능 추가 (roomllist 길게 누르기)

+++ ChatClient/Assets/Scripts/Scenes/MainScene.cs
@@ -138,7 +138,7 @@ public class MainScene : BaseScene
         UI.AddRoomList(roomInfo);
     }
 
-    public void RemoveRoomUI(ulong roomNumber)
+    public void RemoveRoomUI(uint roomNumber)
     {
         UI.RemoveRoom(roomNumber);
     }
